requirements.txt
uzu-accounts-app==1.1.9
python-dotenv==0.15.0